pyproject.toml
[project]
name = "investment-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.5.3",
    "channels>=4.2.2",
    "daphne>=4.2.0",
    "djangorestframework-simplejwt>=5.5.0",
    "graphene-django>=3.2.3",
    "numpy>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pyotp>=2.9.0",
    "python-decouple>=3.8",
    "reportlab>=4.4.1",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
]
